Addressed all review comments.
