+++ MyJobApplication/res/values/strings.xml
@@ -6,7 +6,7 @@
     <!-- Sender Fragment strings -->
     <string name="first_name_label">First Name:</string>
     <string name="last_name_label">Last Name:</string>
-    <string name="email_id_label">EmailID:</string>
+    <string name="email_address_label">Email Address:</string>
     <string name="company_label">Company:</string>
     <string name="location_label">Location:</string>
     <string name="add_button_text">ADD</string>
